more 64 bit compiler warnings removal. Mainly items that will be of no operational impact for FreeWRL.

+++ freex3d/src/lib/input/EAIHeaders.h
@@ -59,7 +59,7 @@ void EAI_RW(char *bufptr);
 
 /* more function prototypes to avoid implicit declarations */
 void Parser_deleteParserForScanStringValueToMem(void);			/* from EAI_C_CommonFunctions.c */
-void Parser_scanStringValueToMem(struct X3D_Node *node, int coffset, int ctype, char *value, int isXML);
+void Parser_scanStringValueToMem(struct X3D_Node *node, size_t coffset, int ctype, char *value, int isXML);
 									/* from EAI_C_CommonFunctions.c */
 int returnRoutingElementLength(int);					/* from EAI_C_CommonFunctions.c */
 void createLoadURL(char *);						/* from EAIEventsIn.c */
